fix: wrong claim button
If a vesting is not active yet, the call to calculateVestedAmounts reverts/throws in the javascript code. Since we don’t catch this subsequent calls to check the token paused state are not done and the UI ends up in a wrong state.

With this change we catch the error and the UI displays the correct info that the vesting cannot be transferred yet.